extract driver_tracing.rs with RequestSpan
